Refactored zendesk ticket creation into its own function
